# Optimize module scoped variables