## Problem
Today when we get or set set a symbol at the module level we defer much of the processing until runtime, via hash lookups of symbols. We already know the module level symbols both in running scripts as well as in the repl so this presents inefficiency. 

## Solution
Instead of performing a hash lookup we perform a simple array index which results in significant speedups in many workloads. This is implemented by a handful of new instructions

* **LoadGlobal**: which loads an instruction from the global module into the current module
* **DeclareModSym**: which pushes a new symbol slot onto the current module with a provided name
* **GetModSym**: which pushes a symbol from the module onto the stack
* **SetModSym** which sets a module symbol at a given slot

### Declaring Symbols
When have three strategies for handling module level symbols

The symbol is found in the global module: 
1. We declare the symbol
2. We load the symbol from the global module
3. We set the symbol into it's slot in the module

The symbol is defined at the module level:
1. We declare the symbol
2. Code is executed until the symbol
3. We set the symbol into it's slot in the module

The symbol was already defined in a repl context
1. We track which slot the symbol sits

In cases all cases when the module is encountered we will load or set the variable using `GetModSym` and `SetModSym`

### Benchmarking

Criterion overall finds each benchmark is either a performance increase or within noise 

```
run/binary_trees        time:   [728.47 µs 730.87 µs 732.93 µs]
                        change: [-2.9221% -2.1459% -1.4102%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) low severe

run/channel             time:   [5.6551 ms 5.6618 ms 5.6688 ms]
                        change: [-2.3289% -2.1819% -2.0287%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  10 (10.00%) high mild
  1 (1.00%) high severe

Benchmarking run/equality: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.9s, enable flat sampling, or reduce sample count to 60.
run/equality            time:   [1.2792 ms 1.2885 ms 1.2982 ms]
                        change: [-3.5379% -2.6607% -1.8850%] (p = 0.00 < 0.05)
                        Performance has improved.

run/fib                 time:   [630.20 µs 631.44 µs 632.78 µs]
                        change: [-1.9341% -1.6175% -1.3233%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  1 (1.00%) high severe

Benchmarking run/fibers: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.7s, enable flat sampling, or reduce sample count to 60.
run/fibers              time:   [1.0522 ms 1.0554 ms 1.0582 ms]
                        change: [-4.3729% -2.9635% -1.5514%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  5 (5.00%) low severe
  13 (13.00%) low mild

run/invocation          time:   [538.56 µs 538.92 µs 539.31 µs]
                        change: [-6.0532% -5.8052% -5.5511%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe

run/instantiation       time:   [903.19 µs 905.07 µs 907.06 µs]
                        change: [-8.1946% -6.9788% -5.6213%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low severe
  1 (1.00%) high severe

run/method_call         time:   [746.67 µs 767.81 µs 815.03 µs]
                        change: [-2.2305% +3.5615% +14.898%] (p = 0.68 > 0.05)
                        No change in performance detected.
Found 17 outliers among 100 measurements (17.00%)
  9 (9.00%) low severe
  3 (3.00%) low mild
  2 (2.00%) high mild
  3 (3.00%) high severe

run/properties          time:   [607.57 µs 608.05 µs 608.57 µs]
                        change: [-6.2962% -6.0107% -5.7019%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

run/trees               time:   [664.98 µs 669.76 µs 674.11 µs]
                        change: [-0.5420% +0.2400% +1.0694%] (p = 0.56 > 0.05)
                        No change in performance detected.

run/zoo                 time:   [468.77 µs 468.92 µs 469.07 µs]
                        change: [-5.5767% -5.2646% -4.9491%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

run/string              time:   [641.03 µs 642.49 µs 643.87 µs]
                        change: [+0.0077% +0.2789% +0.5315%] (p = 0.05 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe

run/fluent              time:   [930.05 µs 931.83 µs 933.76 µs]
                        change: [+0.4662% +0.7025% +0.9447%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```